Do not provide test-helper feature
"test-helper.el" isn't a library intended to be loaded with `require`.
Instead it loaded with `load`.  Because of that, it is confusing to
provide a feature using `provide` anyway.  And furthermore, because
around 70 other packages come with a file named "test-helper.el" and
several dozen of those still provide the feature `test-helper`, this
also leads to a, admittedly somewhat theoretical, feature conflict.

This is also discussed at https://github.com/rejeep/ert-runner.el/issues/38.